use a core package for eveything related to config
