+++ sage/database/core/graph.py
@@ -1,69 +1,7 @@
-# datasets.py
-# Author: Thomas MINIER - MIT License 2017-2018
-from yaml import load
-from sage.database.import_manager import builtin_backends
-from math import inf
-from urllib.parse import quote_plus
+# graph.py
+# Author: Thomas MINIER - MIT License 2017-2019
 import logging
-
-
-def load_config(config_file="config.yaml"):
-    """Load config file to initialize fragment factories.
-    A config file is a Python file, loaded as a module.
-
-    Example config file:
-    # config.yaml
-    name: My LDF server
-    maintainer: chuck Norris <me@gmail.com>
-
-    datasets:
-    -
-        name: DBpedia-2016-04
-        description: DBpedia dataset, version 2016-04
-        backend: hdt-file
-        file: /home/chuck-norris/dbpedia-2016-04.hdt
-    -
-        name: Chuck-Norris-facts
-        description: Best Chuck Norris facts ever
-        backend: rdf-file
-        format: nt
-        file: /home/chuck-norris/facts.nt
-    """
-    config = load(open(config_file))
-    # available backends (populated with sage's native backends)
-    backends = builtin_backends()
-    # build custom backend (if there is some)
-    if 'backends' in config and len(config['backends']) > 0:
-        for b in config['backends']:
-            if 'name' not in b or 'path' not in b or 'connector' not in b or 'required' not in b:
-                raise SyntaxError('Invalid backend declared. Each custom backend must be declared with properties "name", "path", "connector" and "required"')
-            backends[b['name']] = import_backend(b['name'], b['path'], b['connector'], b['required'])
-    # set time quantum
-    if 'quota' in config:
-        if config['quota'] == 'inf':
-            logging.warning("You are using SaGe with an infinite time quantum. Be sure to configure the Worker timeout of Gunicorn accordingly, otherwise long-running queries might be terminated.")
-            quota = inf
-        else:
-            quota = config['quota']
-    else:
-        quota = 75
-    config['quota'] = quota
-    # set page size, i.e. the number of triples per page
-    max_results = config['max_results'] if 'max_results' in config else inf
-    # recopy default config. options in all config objects when they are missing
-    for c in config["datasets"]:
-        if 'quota' not in c:
-            c['quota'] = quota
-        if 'max_results' not in c:
-            c['max_results'] = max_results
-        if 'publish' not in c:
-            c['publish'] = False
-        if 'queries' not in c:
-            c['queries'] = []
-    # build RDF graphs
-    graphs = {c["name"]: Graph(c, backends) for c in config["datasets"]}
-    return (config, graphs, backends)
-
+from urllib.parse import quote_plus
 
 class Graph(object):
     """A RDF Graph with a dedicated backend"""
@@ -164,55 +102,3 @@ class Graph(object):
             if query['@id'] == q_id:
                 return query
         return None
-
-
-class Dataset(object):
-    """A collection of RDF graphs"""
-
-    def __init__(self, config_file):
-        super(Dataset, self).__init__()
-        self._config_file = config_file
-        (self._config, self._datasets, self._backends) = load_config(self._config_file)
-        if "long_description" in self._config:
-            with open(self._config["long_description"], "r") as file:
-                self._long_description = file.read()
-        else:
-            self._long_description = ""
-
-    @property
-    def name(self):
-        return self._config["name"] if "name" in self._config else None
-
-    @property
-    def default_query(self):
-        default = {
-            "name": "",
-            "value": ""
-        }
-        return self._config["default_query"] if "default_query" in self._config else default
-
-    @property
-    def long_description(self):
-        return self._long_description
-
-    @property
-    def public_url(self):
-        return self._config["public_url"] if "public_url" in self._config else None
-
-    @property
-    def maintainer(self):
-        return self._config["maintainer"] if "maintainer" in self._config else None
-
-    def describe(self, url):
-        """Gives a generator over dataset descriptions"""
-        for name, dataset in self._datasets.items():
-            yield dataset.describe(url)
-
-    def get_graph(self, dataset_name):
-        """Get a dataset given its name"""
-        dataset = self._datasets[dataset_name] if dataset_name in self._datasets else None
-        return dataset
-
-    def has_graph(self, dataset_name):
-        """Test fi a graph exist"""
-        return dataset_name in self._datasets
