+++ sage/database/core/rdf_config.py
@@ -1,0 +1,2 @@
+# rdf_config.py
+# Author: Thomas MINIER - MIT License 2017-2019
